# OpenAPI.yaml
openapi: 3.0.0
info:
  title: store
  version: '1.0'
  description: A simple store
  contact:
    name: SecuritEase Dev
    url: 'https://www.securitease.com'
    email: internal@securitease.com
  termsOfService: 'https://www.securitease.com'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.htm'
servers:
  - url: 'http://localhost:8080'
    description: dev
paths:
  /order:
    get:
      summary: Get Orders
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
      operationId: getOrders
      description: Get all orders
    post:
      summary: Create Order
      tags:
        - store
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: The order references no customer, or contains no products
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: The referenced customer or one of the products does not exist
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      operationId: createOrder
      description: >-
        Create a new order. The customer and products are referenced by id and must already
        exist; an order must contain at least one product.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - description
                - customer
                - products
              properties:
                description:
                  type: string
                customer:
                  $ref: '#/components/schemas/Reference'
                products:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/Reference'
  '/order/{id}':
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get Order
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: No order exists with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      operationId: getOrderById
      description: Get a single order by id
  /customer:
    get:
      summary: Get Customers
      tags:
        - store
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: >-
            Case-insensitive substring match against the customer name. Omit to return all
            customers. LIKE wildcards in the value are treated as literal characters.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
      operationId: getCustomers
      description: Get all customers, or search them by name
    post:
      summary: Create Customer
      tags:
        - store
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
      operationId: createCustomer
      description: Create a new customer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
  /products:
    get:
      summary: Get Products
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
      operationId: getProducts
      description: Get all products
    post:
      summary: Create Product
      tags:
        - store
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
      operationId: createProduct
      description: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - description
              properties:
                description:
                  type: string
  '/products/{id}':
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get Product
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: No product exists with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      operationId: getProductById
      description: Get a single product by id
components:
  schemas:
    Reference:
      title: Reference
      type: object
      description: A reference to an existing entity by id.
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
    Order:
      title: Order
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
        products:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              description:
                type: string
    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              description:
                type: string
    Product:
      title: Product
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        orders:
          type: array
          description: Ids of the orders containing this product.
          items:
            type: integer
            format: int64
    ProblemDetail:
      title: ProblemDetail
      type: object
      description: RFC 7807 problem detail.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
tags:
  - name: store
    description: store operations