# src/main/java/com/example/store/controller/CustomerController.java
package com.example.store.controller;

import com.example.store.dto.CustomerDTO;
import com.example.store.entity.Customer;
import com.example.store.mapper.CustomerMapper;
import com.example.store.service.CustomerService;

import lombok.RequiredArgsConstructor;

import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/customer")
@RequiredArgsConstructor
public class CustomerController {

    private final CustomerService customerService;
    private final CustomerMapper customerMapper;

    /**
     * @param query optional; when supplied, returns customers whose name contains it (case-insensitive). Omitting it
     *     returns all customers.
     */
    @GetMapping
    public List<CustomerDTO> getCustomers(@RequestParam(required = false) String query) {
        return customerMapper.customersToCustomerDTOs(customerService.search(query));
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public CustomerDTO createCustomer(@RequestBody Customer customer) {
        return customerMapper.customerToCustomerDTO(customerService.create(customer));
    }
}